// web/src/gen/ChatMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Author } from "./Author";
import type { MessageId } from "./MessageId";

export type ChatMessage = { id: MessageId, author: Author, body: string, createdAt: number, deliveredAt: number | null, };