// web/src/gen/ClientMsg.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Anchor } from "./Anchor";
import type { MessageId } from "./MessageId";
import type { ThreadId } from "./ThreadId";

export type ClientMsg = { "type": "comment", threadId: ThreadId, messageId: MessageId, anchor: Anchor, body: string, } | { "type": "reply", threadId: ThreadId, messageId: MessageId, body: string, } | { "type": "resolve", threadId: ThreadId, resolved: boolean, } | { "type": "drafting", drafting: boolean, } | { "type": "chat", messageId: MessageId, body: string, } | { "type": "read", seq: number, };